+++ Sources/Core/VesselImpl.h
@@ -33,7 +33,7 @@ namespace UACS
 
 			double GetTotalCargoMass() override;
 
-			int GetAvailableCargoCount() override;
+			size_t GetAvailableCargoCount() override;
 
 			std::string_view GetAvailableCargoName(size_t) override;
